IrodsFileSystemConnection : Ensure temporary directory is well set before deleting it

+++ opensilex-fs/src/main/java/org/opensilex/fs/irods/IrodsFileSystemConnection.java
@@ -59,7 +59,9 @@ public class IrodsFileSystemConnection extends BaseService implements Service, F
 
     @Override
     public void shutdown() throws Exception {
-        Files.deleteIfExists(tmpDirectory);
+        if(tmpDirectory != null){
+            Files.deleteIfExists(tmpDirectory);
+        }
     }
 
     private void irodsCommand(String... args) throws IOException {
